requirements.txt
beautifulsoup4==4.6.0
Django==3.2.16
django-appconf==1.0.5
django-blog-zinnia==0.19
django-ckeditor==5.4.0
django-contrib-comments==1.8.0
django-imagekit==4.1.0
django-js-asset==1.0.0
django-tagging==0.4.6
django-tastypie==0.14.4
django-xmlrpc==0.1.8
mots-vides==2015.5.11
olefile==0.44
pilkit==2.0
Pillow==9.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
social-auth-app-django==3.1.0
zinnia-wysiwyg-ckeditor==1.3